@@ -4,7 +4,7 @@ version: 1.6.4
 homepage: https://github.com/peng8350/flutter_pulltorefresh
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=0.1.4 <2.0.0"
 
 dependencies:
